+++ Lawn/System/TypingCheck.cpp
@@ -6,7 +6,7 @@ TypingCheck::TypingCheck(const std::string &thePhrase) { SetPhrase(thePhrase); }
 
 // 0x51C4D0
 void TypingCheck::SetPhrase(const std::string &thePhrase) {
-    for (int i = 0; i < thePhrase.size(); i++)
+    for (size_t i = 0; i < thePhrase.size(); i++)
         AddChar(thePhrase[i]);
 }
 
@@ -30,7 +30,7 @@ bool TypingCheck::Check() {
 // 0x51C5A0
 bool TypingCheck::Check(Sexy::KeyCode theKeyCode) {
     mRecentTyping.append(1, (char)theKeyCode);
-    int aLength = mPhrase.size();
+    size_t aLength = mPhrase.size();
     if (aLength == 0) return false;
 
     if (mRecentTyping.size() > aLength) mRecentTyping = mRecentTyping.substr(1, aLength);
