+++ Sexy.TodLib/TodStringFile.cpp
@@ -78,7 +78,7 @@ bool TodStringListReadName(const char *&thePtr, std::string &theName) {
 
 // 0x518F60
 void TodStringRemoveReturnChars(std::string &theString) {
-    for (int i = 0; i < theString.size();) {
+    for (size_t i = 0; i < theString.size();) {
         if (theString[i] == '\r') theString.replace(i, 1, "", 0); // 原版中此处的“1”和“""”已内联至函数内部
         else i++;
     }
@@ -231,7 +231,7 @@ int TodWriteString(
         }
     }
 
-    if (theLength < 0 || theOffset + theLength > theString.size()) theLength = theString.size();
+    if (theLength < 0 || theOffset + theLength > (int)theString.size()) theLength = theString.size();
     else theLength = theOffset + theLength; // 将 theLength 更改为子串结束位置
 
     SexyString aString;
@@ -309,8 +309,8 @@ int TodDrawStringWrappedHelper(
     int aYOffset = theFont->GetAscent() - theFont->GetAscentPadding();
     int aLineSpacing = theFont->GetLineSpacing() + aCurrentFormat.mLineSpacingOffset;
     SexyString aCurString;
-    int aLineFeedPos = 0;
-    int aCurPos = 0;
+    size_t aLineFeedPos = 0;
+    size_t aCurPos = 0;
     int aCurWidth = 0;
     SexyChar aCurChar = '\0';
     SexyChar aPrevChar = '\0';
